Resource Based Authorization and custom authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestaurant/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,6 +1,9 @@
 using AutoMapper;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
+using Restaurants.Application.Interfaces;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
@@ -9,7 +12,8 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDishes.DeleteDishesForRe
 
 public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService authorizationService) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -17,6 +21,8 @@ public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRest
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
         if (restaurant == null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
 
+        if(!authorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbidException();
        
         await dishesRepository.Delete(restaurant.Dishes);
