+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,5 +1,8 @@
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
+using Restaurants.Application.Interfaces;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Repositories;
@@ -7,7 +10,8 @@ using Restaurants.Domain.Repositories;
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
 public class DeleteRestaurantCommandHandler (ILogger<DeleteRestaurantCommandHandler> logger,
-    IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+    IRestaurantsRepository restaurantsRepository,
+    IRestaurantAuthorizationService authorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -17,6 +21,10 @@ public class DeleteRestaurantCommandHandler (ILogger<DeleteRestaurantCommandHand
         if(restaurant == null)
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
 
+        if(!authorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbidException();
+            
+        
         await restaurantsRepository.Delete(restaurant);
         
     }
